Refactor import path for NumberTeamSelector component
